+++ readme.txt
@@ -3,9 +3,9 @@ Contributors: realloc
 Donate link: http://www.greenpeace.org/international/en/supportus/
 Tags: multisite, recent posts, taxonomy, category, tag, widget
 Requires at least: 4.6
-Tested up to: 5.2
-Stable tag: 1.1
-Requires PHP: 5.6
+Tested up to: 5.5
+Stable tag: 1.2
+Requires PHP: 7.0
 License: GPLv2 or later
 License URI: http://www.gnu.org/licenses/gpl-2.0.html
